+++ src/ai/backend/client/cli/__main__.py
@@ -1,4 +1,4 @@
-from . import run_main
+from . import main
 
 
-run_main()
+main()