+++ src/ai/backend/client/cli/dotfile.py
@@ -3,12 +3,12 @@ import sys
 import click
 from tabulate import tabulate
 
-from . import AliasGroup, main
+from . import main
 from .pretty import print_info, print_warn, print_error
 from ..session import Session
 
 
-@main.group(cls=AliasGroup)
+@main.group()
 def dotfile():
     '''Provides dotfile operations.'''